+++ magda-web-client/src/Components/SearchFacets/SearchFacets.js
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
 import { config } from "../../config";
-import { Small, Medium } from "../../UI/Responsive";
-import downArrowDark from "../../assets/downArrowDark.svg";
+import { Medium } from "../../UI/Responsive";
+// import downArrowDark from "../../assets/downArrowDark.svg";
 import ClearAllButtom from "./ClearAllButton";
 
 import "./SearchFacets.css";
@@ -13,7 +13,7 @@ class SearchFacets extends Component {
         this.toggleFacet = this.toggleFacet.bind(this);
         this.closeFacetWithKeyBoard = this.closeFacetWithKeyBoard.bind(this);
         this.onToggleFacetOnMobile = this.onToggleFacetOnMobile.bind(this);
-        this.renderMobile = this.renderMobile.bind(this);
+        // this.renderMobile = this.renderMobile.bind(this);
         this.renderDesktop = this.renderDesktop.bind(this);
     }
 
@@ -87,47 +87,10 @@ class SearchFacets extends Component {
         );
     }
 
-    renderMobile() {
-        return (
-            <div className="search-facets-mobile">
-                <button
-                    className="filter-toggle-button au-btn"
-                    onClick={this.onToggleFacetOnMobile}
-                >
-                    Filters <img src={downArrowDark} alt="open filter" />
-                </button>
-                {this.state.showFilterOnMobile &&
-                    config.facets.map(c => (
-                        <div
-                            className="search-facet"
-                            key={c.id}
-                            onClick={ev => ev.stopPropagation()}
-                        >
-                            <c.component
-                                updateQuery={this.props.updateQuery}
-                                location={this.props.location}
-                                title={c.id}
-                                isOpen={this.state.openFacet === c.id}
-                                toggleFacet={this.toggleFacet.bind(this, c.id)}
-                                closeFacet={this.closeFacet.bind(this, c.id)}
-                            />
-                        </div>
-                    ))}
-                {this.state.showFilterOnMobile && (
-                    <ClearAllButtom key={"clear-all-button"} />
-                )}
-                {this.state.openFacet && (
-                    <div className="mobile-facet-background" />
-                )}
-            </div>
-        );
-    }
-
     render() {
         return (
             <div>
                 <Medium>{this.renderDesktop()}</Medium>
-                <Small>{this.renderMobile()}</Small>
             </div>
         );
     }